refactor: Reorganise components
